nn_training/requirements.txt
torch
torchvision
pyyaml
numpy
python-chess
tqdm
